requirements.txt
contourpy==1.3.3
cycler==0.12.1
fonttools==4.62.1
glib==1.0.0
gpib-ctypes==0.3.0
ifaddr==0.2.0
kiwisolver==1.5.0
matplotlib==3.10.8
numpy==2.4.4
packaging==26.0
pillow==12.2.0
psutil==7.2.2
psutils==3.3.14
puremagic==2.1.1
pyparsing==3.3.2
pypdf==6.9.2
python-dateutil==2.9.0.post0
python-usbtmc==0.8
pyusb==1.3.1
pyvicp==1.1.0
PyVISA==1.16.2
PyVISA-py==0.8.1
six==1.17.0
typing_extensions==4.15.0
zeroconf==0.148.0
